+++ src/grammar/GrammarConstants.ts
@@ -13,6 +13,7 @@ enum GrammarConstants {
   grammar = "grammar",
   extensions = "extensions",
   version = "version",
+  keywordOrder = "keywordOrder",
   keyword = "keyword",
   wordType = "wordType",
   abstract = "abstract",